feat: make tmux the first backend behind the Multiplexer Protocol (Task 2)
